@@ -2,7 +2,6 @@ package restapi
 
 import (
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 )
@@ -188,7 +187,7 @@ func getTestServer() *httptest.Server {
 			}
 
 			// basic check to see if a POST has a request body
-			bodyBytes, err := ioutil.ReadAll(req.Body)
+			bodyBytes, err := io.ReadAll(req.Body)
 			if err != nil {
 				status = http.StatusBadRequest
 			}
